Switched the counter variable of all of the loops in the update() method to use shorts  instead of ints as the highest counter reached can never be larger than Short.MAX_VALUE.

+++ Particles/src/effect/RainbowSnow.java
@@ -63,7 +63,7 @@ public class RainbowSnow extends Effect {
         // update() method of the Effect class because it's divisible
         // by 2.
 		if(counter == 10) {
-			for(int i=0;i<TOTAL_PARTICLES/100;i++) { newParticle(); }
+			for(short i=0;i<TOTAL_PARTICLES/100;i++) { newParticle(); }
 			counter = 0;
 		} else {
 			counter++;
